backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic-settings==2.4.0
pdfplumber==0.11.4
pillow==10.4.0
pytesseract==0.3.10
openai==1.55.0
sentence-transformers==3.1.1
chromadb==0.5.5
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
openai-whisper==20231117
whisper-timestamped==1.14.0
numpy==2.1.3
ijson==3.3.0
pinecone-client==4.1.2
qdrant-client==1.11.1
websockets==12.0

